wasserstoff/AiTask/requirements.txt
ipykernel
streamlit
python-dotenv
langchain
langchain-community
ollama
fastapi
uvicorn
langserve
sse_starlette
bs4
pypdf
chromadb
wikipedia
arxiv
langchainhub
sentence-transformers
faiss-cpu
flask